backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.6.1
pillow==10.3.0
numpy==1.26.4
gspread==6.1.0
google-auth==2.29.0
watchfiles==0.21.0
torch==2.8.0
torchvision==0.23.0b 
